Add files via upload
Adding an auto clicker was very difficult because mit app inventor always crashed when it tried to run the auto clicker. Another thing was difficult was creating another screen because the variables dont save on the other screen. Lastly, desinging the actual interface was difficult because I was too lazy.